fix test_write_df() and test_write_tb_events()
