+++ .github/workflows/publish.yml
@@ -13,9 +13,15 @@ jobs:
       - name: Check out repo
         uses: actions/checkout@v2
 
+      - name: Set up python
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.8
+          cache: pip
+          cache-dependency-path: setup.cfg
+
       - name: Install dependencies
-        run: |
-          pip install .[test]
+        run: pip install .[test]
 
       - name: Run tests
         # Only publish to PyPI if tests pass.
